+++ back/src/main/java/com/bubble/giju/domain/cart/dto/request/AddToCartRequestDto.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.Bag;
 
 @Getter
 @NoArgsConstructor
