+++ back/src/main/java/com/bubble/giju/domain/cart/dto/request/CartSelectedRequestDto.java
@@ -1,6 +1,5 @@
 package com.bubble.giju.domain.cart.dto.request;
 
-import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
