+++ back/src/main/resources/templates/payment.html
@@ -25,7 +25,7 @@
         const coupon = document.getElementById("coupon-box");
         // ------  Initialize widget ------
 
-        const clientKey = "클라이언트 key입력";
+        const clientKey = "test_gck_docs_Ovk5rk1EwkEbP0W43n07xlzm";
         const tossPayments = TossPayments(clientKey);
         // Logged in customers
         const customerKey = "ne93IVS3YN7kaA_Ag7Dts";
@@ -38,7 +38,7 @@
         // ------ Set payment amount ------
         await widgets.setAmount({
             currency: "KRW",
-            value: 3500,
+            value: 6500,
         });
 
         await Promise.all([
@@ -64,7 +64,7 @@
 
             await widgets.setAmount({
                 currency: "KRW",
-                value: 3500,
+                value: 6500,
             });
         });
 
@@ -72,7 +72,7 @@
         // Save the amount and orderId to prevent malicious attempts to change them
         button.addEventListener("click", async function () {
             await widgets.requestPayment({
-                orderId: "ORDER_5000_i1j2k3l4-m5n6-vwxy-ijkl-zabcdefghijk",
+                orderId: "ORDER_60_h0i1j2k3-l4m5-uvwx-hijk-yzabcdefghig",
                 orderName: "토스 티셔츠 외 2건",
                 successUrl: "http://localhost:8080/api/payment/success",
                 failUrl: "http://localhost:8080/api/payment/fail",
